Add parameters to use it for the patch
Also push the target include into the source toolchain (ouch).

+++ ubuntu16.04/build-toolchain.sh
@@ -16,7 +16,11 @@ SWIFT_VERSION=${SWIFT_VERSION:=5.2}
 TARGET_ARCH=${TARGET_ARCH:=x86_64}
 TARGET_PLATFORM=${TARGET_PLATFORM:=ubuntu16.04}
 CROSS_TOOLCHAIN_NAME=${CROSS_TOOLCHAIN_NAME:=swift-${SWIFT_VERSION}-${TARGET_PLATFORM}.xtoolchain}
-HOST_PLATFORM=x86_64
+HOST_PLATFORM=${HOST_PLATFORM:=x86_64}
+
+# must be specified, absolute URL:
+#   e.g. /usr/local/lib/swift/dst/${TARGET_ARCH}-unknown-linux}
+INSTALL_PREFIX=${INSTALL_PREFIX:=${BUILD_DIR}}
 
 # brew install swiftxcode/swiftxcode/swift-xctoolchain-5.2
 # brew install swiftxcode/swiftxcode/clang-llvm-bin-8
@@ -143,12 +147,17 @@ cp -ac "$HOST_X_LLD" "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/${xc_tc_name}/usr/bin
 
 # swift.xctoolchain/usr/lib/swift_static/linux
 # swift.xctoolchain/usr/lib/swift/linux
-echo "Coping in headers/libs from target Swift toolchain: ${linux_swift_pkg} ..."
 (
-# FIXME: just tar xvf over the right location
-tmp=$(mktemp -d "${BUILD_DIR}/tmp_pkgs_XXXXXX")
-tar -C "$tmp" --strip-components 1 -xf "$linux_swift_pkg"
-UNPACKED_LINUX_TC="$tmp"
+if [[ "x$linux_swift_pkg" != "x" ]]; then
+  echo "Coping in headers/libs from target Swift toolchain: ${linux_swift_pkg} ..."
+  tmp=$(mktemp -d "${BUILD_DIR}/tmp_pkgs_XXXXXX")
+  tar -C "$tmp" --strip-components 1 -xf "$linux_swift_pkg"
+  UNPACKED_LINUX_TC="$tmp"
+else
+  echo "Coping in headers/libs from target Swift toolchain (inline) ..."
+  UNPACKED_LINUX_TC="."
+  file -d usr
+fi
 
 # TBD: This might not be necessary when we use `-resource-dir` to point the
 #      compiler to the actual Ubuntu SDK (it defaults to the host)!
@@ -171,18 +180,29 @@ cp -ac "${UNPACKED_LINUX_TC}/usr/lib/swift/dispatch" \
        "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/lib/swift/dispatch"
 cp -ac "${UNPACKED_LINUX_TC}/usr/lib/swift/os" \
        "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/lib/swift/os"
-mv "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/lib/swift/shims" \
-   "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/lib/swift/shims.org"
+# includes twice when leaving the .org inside ...
+#mv "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/lib/swift/shims" \
+#   "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/lib/swift/shims.org"
+rm -rf "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/lib/swift/shims"
 cp -ac "${UNPACKED_LINUX_TC}/usr/lib/swift/shims" \
        "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/lib/swift/shims"
 
+# Another Hack which works around the wrong includes being picked up.
+# We essentially morph the Host toolchain within into a target one ...
+# Maybe the above should just do the same ...
+mv ${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/include \
+   ${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/include.org
+ln -s ../../$linux_sdk_name/usr/include \
+      ${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/include
 
 
 echo "  .. Linux Swift libs/mods into target toolchain ..."
 cp -ac "${UNPACKED_LINUX_TC}/usr/lib/swift"        "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$linux_sdk_name/usr/lib/swift"
 cp -ac "${UNPACKED_LINUX_TC}/usr/lib/swift_static" "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$linux_sdk_name/usr/lib/swift_static"
 
-rm -rf "$tmp"
+if [[ "x$linux_swift_pkg" != "x" ]]; then
+  rm -rf "$tmp"
+fi
 echo "  ok."
 )
 
@@ -210,16 +230,16 @@ fix_glibc_modulemap "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$linux_sdk_name/usr/li
 cat > "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/destination.json" <<EOF
 {
     "version": 1,
-    "sdk": "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$linux_sdk_name",
-    "toolchain-bin-dir": "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/bin",
+    "sdk": "${INSTALL_PREFIX}/${CROSS_TOOLCHAIN_NAME}/$linux_sdk_name",
+    "toolchain-bin-dir": "${INSTALL_PREFIX}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/bin",
     "target": "x86_64-unknown-linux",
     "extra-cc-flags": [
         "-fPIC"
     ],
     "extra-swiftc-flags": [
         "-use-ld=lld", 
-        "-tools-directory", "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/bin",
-        "-sdk", "${BUILD_DIR}/${CROSS_TOOLCHAIN_NAME}/$linux_sdk_name"
+        "-tools-directory", "${INSTALL_PREFIX}/${CROSS_TOOLCHAIN_NAME}/$xc_tc_name/usr/bin",
+        "-sdk", "${INSTALL_PREFIX}/${CROSS_TOOLCHAIN_NAME}/$linux_sdk_name"
     ],
     "extra-cpp-flags": [
         "-lstdc++"
